rename controllers to services and add readme stuff
