+++ README.md
@@ -2,3 +2,19 @@
 my hapi bootstrap for new projects
 
 just a shell project for my node.js hapi setups
+
+##requirements
+* NVM
+* Node 4
+
+## to run unit tests
+`gulp`
+
+## to run integration tests
+`gulp integrationTest --env=local`
+
+where env = the environment you want to run against (see config.js)
+
+## to run the server
+
+`node index.js --env=local`
